refactor: simplify fake provider bridge structure
